DRHP_crud_backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
azure-ai-documentintelligence==1.0.0b4
azure-ai-formrecognizer==3.3.3
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.19.0
azure-storage-blob==12.20.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
boto3==1.36.12
botocore==1.36.17
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.16
duckduckgo-search==7.3.2
email-validator==2.2.0
et-xmlfile==2.0.0
fastapi==0.115.8
filelock==3.17.0
forex-python==1.8
frozenlist==1.5.0
fsspec==2025.2.0
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.6.1
instructor==1.7.2
isodate==0.7.2
jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.60.8
lxml==5.3.1
markdown-it-py==3.0.0
markdownify==1.1.0
markupsafe==3.0.2
mdurl==0.1.2
mongoengine==0.29.1
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.1.0
numpy==2.2.2
oauthlib==3.2.2
openai==1.61.1
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdfminer-six==20231228
pdfplumber==0.11.5
pillow==11.1.0
portalocker==2.10.1
primp==0.12.1
propcache==0.2.1
protobuf==5.29.3
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
PyJWT
pymongo==4.11
pymupdf==1.25.3
pypdf==5.2.0
pypdf2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyyaml==6.0.2
pyzbar==0.1.9
qdrant-client==1.13.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.11.2
setuptools==75.8.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
typer==0.15.1
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
zipp==3.21.0
fastembed
baml-py
openai
